Improve app performce by removing duplicate API calls
