+++ src/client/components/NotificationAlert/reducer.js
@@ -1,12 +1,20 @@
-import { NOTIFICATION_ALERT_COUNT__LOADED } from '../../actions'
+import { REMINDER_SUMMARY__LOADED } from '../../actions'
 
 const initialState = {
   count: 0,
+  investment: {
+    estimated_land_date: 0,
+    no_recent_interaction: 0,
+    outstanding_propositions: 0,
+  },
+  export: {
+    no_recent_interaction: 0,
+  },
 }
 
 export default (state = initialState, { type, result }) => {
   switch (type) {
-    case NOTIFICATION_ALERT_COUNT__LOADED:
+    case REMINDER_SUMMARY__LOADED:
       return {
         ...result,
       }
