+++ src/client/components/NotificationAlert/tasks.js
@@ -1,8 +1,4 @@
 import { apiProxyAxios } from '../Task/utils'
 
-const transformNotificationAlertCount = ({ data }) => ({ count: data.count })
-
-export const fetchNotificationAlertCount = () =>
-  apiProxyAxios
-    .get('/v4/reminder/summary')
-    .then(transformNotificationAlertCount)
+export const fetchReminderSummary = () =>
+  apiProxyAxios.get('/v4/reminder/summary').then(({ data }) => data)
