+++ src/client/components/ReminderSummary/index.jsx
@@ -1,27 +1,92 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
+import styled from 'styled-components'
+import { kebabCase } from 'lodash'
 
-import { ID, TASK_GET_REMINDER_SUMMARY, state2props } from './state'
-import { REMINDER_SUMMARY__LOADED } from '../../actions'
+import { LINK_COLOUR } from 'govuk-colours'
+import { H5 } from '@govuk-react/heading'
+import { FONT_SIZE, SPACING, FONT_WEIGHTS } from '@govuk-react/constants'
 
-import Summary from './Summary'
-import Task from '../Task'
+import urls from '../../../lib/urls'
+import { state2props } from './state'
 
-const ReminderSummary = (data) => (
-  <div data-test="reminder-summary">
-    <Task.Status
-      name={TASK_GET_REMINDER_SUMMARY}
-      id={ID}
-      progressMessage="Loading your reminders"
-      startOnRender={{
-        onSuccessDispatch: REMINDER_SUMMARY__LOADED,
-      }}
-    >
-      {() => <Summary {...data} />}
-    </Task.Status>
-  </div>
-)
+const StyledSubHeading = styled(H5)`
+  font-size: ${FONT_SIZE.SIZE_10};
+  font-weight: ${FONT_WEIGHTS.bold};
+  margin-top: ${SPACING.SCALE_2};
+  margin-bottom: ${SPACING.SCALE_2};
+`
+
+const StyledReminderLink = styled('a')`
+  display: inline;
+  font-size: ${FONT_SIZE.SIZE_16};
+  color: ${LINK_COLOUR};
+`
+
+const StyledList = styled('ul')`
+  list-style-type: disc;
+  padding: 0 0;
+  margin: ${SPACING.SCALE_5} ${SPACING.SCALE_5};
+`
+
+const StyledListItem = styled('li')(() => ({
+  margin: `${SPACING.SCALE_2} 0`,
+}))
+
+const ReminderSummary = ({
+  summary,
+  hasInvestmentFeatureGroup,
+  hasExportFeatureGroup,
+}) => {
+  const showInvestment = hasInvestmentFeatureGroup && !!summary
+  const showExport = hasExportFeatureGroup && !!summary
+  return (
+    <div data-test="reminder-summary">
+      {showInvestment && (
+        <>
+          <StyledSubHeading data-test="investment-heading">
+            Investment
+          </StyledSubHeading>
+          <StyledList>
+            {summary.investment.map((reminder) => (
+              <StyledListItem
+                key={reminder.name}
+                data-test={`investment-${kebabCase(reminder.name)}`}
+              >
+                <StyledReminderLink href={reminder.url}>
+                  {reminder.name}
+                </StyledReminderLink>
+                &nbsp;({reminder.count})
+              </StyledListItem>
+            ))}
+          </StyledList>
+        </>
+      )}
+      {showExport && (
+        <>
+          <StyledSubHeading data-test="export-heading">Export</StyledSubHeading>
+          <StyledList>
+            {summary.export.map((reminder) => (
+              <StyledListItem
+                key={reminder.name}
+                data-test={`export-${kebabCase(reminder.name)}`}
+              >
+                <StyledReminderLink href={reminder.url}>
+                  {reminder.name}
+                </StyledReminderLink>
+                &nbsp;({reminder.count})
+              </StyledListItem>
+            ))}
+          </StyledList>
+        </>
+      )}
+      <StyledReminderLink href={urls.reminders.settings.index()}>
+        Settings: reminders and email notifications
+      </StyledReminderLink>
+    </div>
+  )
+}
 
 const reminderType = PropTypes.arrayOf(
   PropTypes.shape({
@@ -32,8 +97,13 @@ const reminderType = PropTypes.arrayOf(
 )
 
 ReminderSummary.propTypes = {
-  count: PropTypes.number,
-  investment: reminderType,
+  summary: PropTypes.shape({
+    count: PropTypes.number,
+    investment: reminderType,
+    export: reminderType,
+  }),
+  hasExportFeatureGroup: PropTypes.bool.isRequired,
+  hasInvestmentFeatureGroup: PropTypes.bool.isRequired,
 }
 
 export default connect(state2props)(ReminderSummary)
