+++ src/client/index.jsx
@@ -214,11 +214,8 @@ import { TASK_DATA_HUB_FEED } from './components/PersonalisedDashboard/state'
 
 import { fetchOutstandingPropositions } from './components/InvestmentReminders/tasks'
 import { TASK_GET_OUTSTANDING_PROPOSITIONS } from './components/InvestmentReminders/state'
-import { fetchReminderSummary } from './components/ReminderSummary/tasks'
-import { TASK_GET_REMINDER_SUMMARY } from './components/ReminderSummary/state'
-
-import { fetchNotificationAlertCount } from './components/NotificationAlert/tasks'
-import { TASK_GET_NOTIFICATION_ALERT_COUNT } from './components/NotificationAlert/state'
+import { fetchReminderSummary } from './components/NotificationAlert/tasks'
+import { TASK_GET_REMINDER_SUMMARY } from './components/NotificationAlert/state'
 
 import { TASK_GET_TYPEAHEAD_OPTIONS } from './components/Typeahead/state'
 
@@ -450,7 +447,6 @@ function App() {
             myInvestmentProjects.fetchMyInvestmentsList,
           [TASK_GET_OUTSTANDING_PROPOSITIONS]: fetchOutstandingPropositions,
           [TASK_GET_REMINDER_SUMMARY]: fetchReminderSummary,
-          [TASK_GET_NOTIFICATION_ALERT_COUNT]: fetchNotificationAlertCount,
           'Large investment profiles filters':
             investmentProfilesTasks.loadFilterOptions,
           [TASK_GET_CONTACTS_LIST]: getContacts,
